+++ packages/app/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tinyhttp/app
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @tinyhttp/router@1.0.5
+
 ## 1.0.5
 
 ### Patch Changes