+++ packages/cookie-parser/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tinyhttp/cookie-parser
 
+## 1.0.6
+
+### Patch Changes
+
+- @tinyhttp/app@1.0.6
+
 ## 1.0.5
 
 ### Patch Changes